refactor: rename submission module to intake and update package names and imports accordingly
