+++ chatbot/pom.xml
@@ -21,7 +21,7 @@
     <dependencies>
         <dependency>
             <groupId>pt.estga</groupId>
-            <artifactId>submission</artifactId>
+            <artifactId>intake</artifactId>
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
         <dependency>
