+++ chatbot/src/main/java/pt/estga/chatbot/features/proposal/handlers/InitialLocationHandler.java
@@ -7,7 +7,7 @@ import pt.estga.chatbot.context.ConversationStateHandler;
 import pt.estga.chatbot.context.HandlerOutcome;
 import pt.estga.chatbot.context.ProposalState;
 import pt.estga.chatbot.models.BotInput;
-import pt.estga.submission.entities.MarkEvidenceSubmission;
+import pt.estga.intake.entities.MarkEvidenceSubmission;
 
 @Component
 public class InitialLocationHandler implements ConversationStateHandler {
