+++ intake/src/main/java/pt/estga/intake/dtos/ProposalSummaryDto.java
@@ -1,0 +1,14 @@
+package pt.estga.intake.dtos;
+
+import pt.estga.intake.enums.SubmissionStatus;
+
+import java.time.Instant;
+
+public record ProposalSummaryDto(
+        Long id,
+        String title,
+        String type,
+        Long photoId,
+        SubmissionStatus status,
+        Instant submittedAt
+) {}
