+++ intake/src/main/java/pt/estga/intake/enums/SubmissionSource.java
@@ -1,0 +1,5 @@
+package pt.estga.intake.enums;
+
+public enum SubmissionSource {
+    WEB_APP, WHATSAPP, API, TELEGRAM_BOT, OTHER
+}
