+++ intake/src/main/java/pt/estga/intake/enums/SubmissionStatus.java
@@ -1,0 +1,8 @@
+package pt.estga.intake.enums;
+
+public enum SubmissionStatus {
+    SUBMITTED,
+    UNDER_REVIEW,
+    MANUALLY_ACCEPTED,
+    MANUALLY_REJECTED
+}
