+++ intake/src/main/java/pt/estga/intake/enums/SubmissionType.java
@@ -1,0 +1,5 @@
+package pt.estga.intake.enums;
+
+public enum SubmissionType {
+    MARK_OCCURRENCE
+}
