+++ intake/src/main/java/pt/estga/intake/events/SubmissionSubmittedEvent.java
@@ -1,0 +1,15 @@
+package pt.estga.intake.events;
+
+import lombok.Getter;
+import org.springframework.context.ApplicationEvent;
+
+@Getter
+public class SubmissionSubmittedEvent extends ApplicationEvent {
+
+    private final Long proposalId;
+
+    public SubmissionSubmittedEvent(Object source, Long proposalId) {
+        super(source);
+        this.proposalId = proposalId;
+    }
+}
