+++ intake/src/main/java/pt/estga/intake/repositories/MarkEvidenceSubmissionRepository.java
@@ -1,0 +1,21 @@
+package pt.estga.intake.repositories;
+
+import org.springframework.data.jpa.repository.EntityGraph;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import org.springframework.stereotype.Repository;
+import pt.estga.intake.entities.MarkEvidenceSubmission;
+
+import java.util.Optional;
+
+@Repository
+public interface MarkEvidenceSubmissionRepository extends JpaRepository<MarkEvidenceSubmission, Long>, JpaSpecificationExecutor<MarkEvidenceSubmission> {
+
+    @EntityGraph(attributePaths = {
+            "originalMediaFile",
+            "submittedBy"
+    })
+    @Override
+    Optional<MarkEvidenceSubmission> findById(Long id);
+
+}
