+++ intake/src/main/java/pt/estga/intake/services/ChatbotSubmissionFacade.java
@@ -1,0 +1,41 @@
+package pt.estga.intake.services;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import pt.estga.intake.entities.MarkEvidenceSubmission;
+import pt.estga.intake.enums.SubmissionSource;
+import pt.estga.user.services.UserService;
+
+import java.io.IOException;
+
+/**
+ * Thin facade that encapsulates chatbot-specific submission orchestration:
+ * - resolves an optional domain user id to a User and attaches it to the submission
+ * - ensures a submission source is set
+ * - delegates the actual persistence and media saving to the application SubmissionService
+ */
+@Component
+@RequiredArgsConstructor
+public class ChatbotSubmissionFacade {
+
+    private final SubmissionService submissionService;
+    private final UserService userService;
+
+    public void submitFromChatbot(
+            MarkEvidenceSubmission submission,
+            byte[] photoData,
+            String photoFilename,
+            Long domainUserId,
+            SubmissionSource source
+    ) throws IOException {
+        if (domainUserId != null) {
+            userService.findById(domainUserId).ifPresent(submission::setSubmittedBy);
+        }
+
+        if (submission.getSubmissionSource() == null) {
+            submission.setSubmissionSource(source != null ? source : SubmissionSource.OTHER);
+        }
+
+        submissionService.submit(submission, photoData, photoFilename);
+    }
+}
