+++ pom.xml
@@ -28,7 +28,7 @@
         <module>file</module>
         <module>user</module>
         <module>shared</module>
-        <module>submission</module>
+        <module>intake</module>
         <module>support</module>
         <module>chatbot</module>
         <module>verification</module>
