--- submission/src/main/java/pt/estga/submission/config/SubmissionAutoConfiguration.java
@@ -1,15 +1,0 @@
-package pt.estga.submission.config;
-
-import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.boot.persistence.autoconfigure.EntityScan;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@AutoConfiguration
-@ComponentScan("pt.estga.submission")
-@EnableJpaRepositories("pt.estga.submission.repositories")
-@EntityScan("pt.estga.submission.entities")
-@EnableCaching
-public class SubmissionAutoConfiguration {
-}
