--- submission/src/main/java/pt/estga/submission/enums/SubmissionSource.java
@@ -1,5 +1,0 @@
-package pt.estga.submission.enums;
-
-public enum SubmissionSource {
-    WEB_APP, WHATSAPP, API, TELEGRAM_BOT, OTHER
-}
