--- submission/src/main/java/pt/estga/submission/enums/SubmissionStatus.java
@@ -1,8 +1,0 @@
-package pt.estga.submission.enums;
-
-public enum SubmissionStatus {
-    SUBMITTED,
-    UNDER_REVIEW,
-    MANUALLY_ACCEPTED,
-    MANUALLY_REJECTED
-}
