--- submission/src/main/java/pt/estga/submission/enums/SubmissionType.java
@@ -1,5 +1,0 @@
-package pt.estga.submission.enums;
-
-public enum SubmissionType {
-    MARK_OCCURRENCE
-}
